Modified how numerical blocks are stored
Each Version class will store data specific to its version.
Before you could add new blocks to the Version class but if an old chunk was found the modified values would not be in that Version class.
Custom values should now be added to TranslationManager.blocks and the old translation methods will look here first to see if the defaults have been overridden