Forced ordering of the version numbers
This forces version_numbers to be ordered. This should make some things a bit easier.
It also forces _version_remap to be populated based on version number.
This means that if two versions have the same data version (*cough* bedrock *cough*) the most recent one will actually be used.
This sort of worked before but the ordering of listdir is arbitrary so it was not guaranteed.